Fix: Resolve problem with item index number
Because of different index for fold change, group rank was being
replaced by that and equated to zero.

+++ libmaven/PeakGroup.h
@@ -99,7 +99,7 @@ class PeakGroup{
         float maxSignalBaselineRatio;
         int goodPeakCount;
         float expectedRtDiff;
-        unsigned int groupRank;
+        float groupRank;
 
         //for sample contrasts  ratio and pvalue
         float changeFoldRatio;
